sct_process_segmentation: lazyloading applied
Former-commit-id: 17697e2cf287eb5e586ebc98f5f886fae02dfdfa [formerly ac50edf939a5a5577e677cbb24150f33218bd1e6 [formerly d3fcb894c6ef00fc720ea971efab9bf279c64223]]
Former-commit-id: ecac03030bf2bf2a7ba5be260169ac36f38cc471
Former-commit-id: 3a80d6d6ee599eb0b3708314e9eb014283a8c6df
